# [WIP][RFC] Edit nn.Linear to take arbitrary in and out feature dimensions