--- a/torch/csrc/api/include/torch/nn/modules/linear.h
+++ b/torch/csrc/api/include/torch/nn/modules/linear.h
@@ -13,10 +13,11 @@ namespace nn {
 /// Options for the `Linear` module.
 struct TORCH_API LinearOptions {
   LinearOptions(int64_t in, int64_t out);
+  LinearOptions(std::vector<int64_t> in, std::vector<int64_t> out);
   /// The number of input features (columns of the input matrix).
-  TORCH_ARG(int64_t, in);
+  TORCH_ARG(std::vector<int64_t>, in);
   /// The number of output features to produce (columns of the output matrix).
-  TORCH_ARG(int64_t, out);
+  TORCH_ARG(std::vector<int64_t>, out);
   /// Whether to learn and add a bias after the linear transformation.
   TORCH_ARG(bool, with_bias) = true;
 };
@@ -24,7 +25,18 @@ struct TORCH_API LinearOptions {
 /// Applies a linear transformation with optional bias.
 class TORCH_API LinearImpl : public Cloneable<LinearImpl> {
  public:
-  LinearImpl(int64_t in, int64_t out) : LinearImpl(LinearOptions(in, out)) {}
+  LinearImpl(int64_t in, int64_t out)
+      : LinearImpl(LinearOptions(in, out)) {}
+
+  LinearImpl(std::vector<int64_t> in, int64_t out)
+      : LinearImpl(LinearOptions(in, {out})) {}
+
+  LinearImpl(int64_t in, std::vector<int64_t>  out)
+      : LinearImpl(LinearOptions({in}, out)) {}
+
+  LinearImpl(std::vector<int64_t> in, std::vector<int64_t> out)
+      : LinearImpl(LinearOptions(in, out)) {}
+
   explicit LinearImpl(LinearOptions options);
 
   void reset() override;
@@ -45,6 +57,9 @@ class TORCH_API LinearImpl : public Cloneable<LinearImpl> {
   /// The learned bias. If `with_bias` is false in the `options`, this tensor is
   /// undefined.
   Tensor bias;
+
+ private:
+   int64_t in_prod_, out_prod_;
 };
 
 /// A `ModuleHolder` subclass for `LinearImpl`.