### Current Behavior

Currently, `nn.Linear` takes two integer parameters `in_features` and `out_features` and creates a weight parameter of size `[out_features, in_features]`, and bias parameter of size `[out_features]`. This API is a little cumbersome if users would like to run higher-dimensional features through a linear layer. Say I have 2D input features of size `[2, 3]`, and provide an input batch of size `[4, 5, 2, 3]`. To use the existing linear layer, I will have to first reshape the the input into size `[4, 5, 6]`, and pass `6` as `in_features` arg for `nn.Linear`. 


### Proposed Behavior

It would be nice if `nn.Linear` can handle arbitrary feature dimensions. For example, if I have input feature size `[2, 3]`, and output feature size `4`, I should be able to do the following:

```python
m = Linear(in_features=[2, 3], out_features=4)
input = torch.rand(4, 5, 2, 3)
output = m(input) # get output of size [4, 5, 4]
```

Let `in_features=[in_1, in_2, ..., in_x]`, `out_features=[out_1, out_2, ..., out_y]`, and `input.size() == [d1, d2, ..., dz]`. The output dimensions should be `[d_1, ..., d_{z-x}, out_1, ..., out_y]`. 

It is allowed to set `in_features` and/or `out_features` to `[]` or `()`, and its behavior will stay consistent with `torch.matmul`.


### Implementation Alternatives

#### Option 1 [implemented in this PR]

Weights and bias dimensions respect `in_features` and `out_features` provided by users, where `weight.size() == [out_1, out_2, ..., out_y, in_1, in_2, ..., in_x]`  and `bias.size() == out_features`. In every `forward()` iteration, `weight` is reshaped to a 2D tensor of size `[prod(out_features), prod(in_features)]`, and `bias` of size `[prod(in_features)]`. Then, inputs are reshaped to `[d_1, ..., d_{z-x}, prod(in_features)]`. `F.linear` function creates an intermediate output of size `[d_1, ..., d_{z-x}, prod(out_features)]`, and are reshaped to `[d_1, ..., d_z, out_1, ..., out_y]`. 

The reason for not keeping weight and bias views as member var is to avoid showing them in `Linear.parameters()` 

Pro: Dimensions shown in `Linear.parameters()` will match users' expectation.
Con: Have to create multiple views in every `forward`.

#### Option 2

In constructor, directly create weight of size `[prod(out_features), prod(in_features)]` and bias of size `[prod(out_features)]`. And only reshape input and output in every `forward()` iteration. 

Pro: Avoid reshaping `weight` and `bias` in `forward()`
Con: Dimensions shown in `Linear.parameters()` won't match `in_features` and `out_features`

#### Option 3
Implement a new `linear` native function which address the dimensions internal, and modify both Python and C++ frontend to call it.

Will edit docs and add comments when we have consensus.

@gchanan @vishwakftw @yf225 

### Benchmark [Python API only]

before

```python
In [2]: l = torch.nn.Linear(10, 10)

In [3]: input = torch.rand(2, 10)

In [4]: %timeit l(input)
26.2 µs ± 5.17 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)

In [5]: %timeit l(input)
25.1 µs ± 1.57 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)

In [6]: %timeit l(input)
25.6 µs ± 1.93 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
```

after

```python
In [2]: l = torch.nn.Linear(10, 10)

In [4]: input = torch.rand(2, 10)

In [5]: %timeit l(input)
24.4 µs ± 383 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)

In [6]: %timeit l(input)
26.2 µs ± 4.25 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)

In [7]: %timeit l(input)
24.8 µs ± 2.37 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
```