@vishwakftw 2019-05-08
I think `view`ing might slightly expensive. On a related note, are there any scenarios where the intermediate tensors cannot be viewed in the current implementation?