@mrshenli 2019-05-08
> On a related note, are there any scenarios where the intermediate tensors cannot be viewed in the current implementation?

@vishwakftw Which tensors?