@vishwakftw 2019-05-08
> @vishwakftw Which tensors?

Few instances where viewing which can possibly throw errors are these
- https://github.com/pytorch/pytorch/pull/20258/files#diff-531f4c06f42260d699f43dabdf741b6dR132
- https://github.com/pytorch/pytorch/pull/20258/files#diff-b3ce6f6da2faf48ee488d334b9833f34R62

Is the output of `F.linear` always guaranteed to be contiguous?

(Sorry about the links; I don't know why permalink isn't available)