@ezyang 2019-05-08
In your PR body, you say:

> The last len(in) dimensions of input data are feature dimensions, and all preceding dimensions are batch dimensions.

What is `in`?