@ezyang 2019-05-08
For my information, what is an example use case for this feature? It's not obvious to me under what circumstances you'd be willing to remove all of the structure of the input dimensions and just matrix multiply it as if it were flat.