@ezyang 2019-05-08
I tried to review the logic change, but I couldn't figure out what the intended semantics were supposed to be. Furthermore, the documentation in the codebase was not updated, so I could not read it to see what the semantics were. And I didn't understand the PR body description (see above).