@gchanan 2019-05-08
> The last len(in) dimensions of input data are feature dimensions, and all preceding dimensions are batch dimensions. For input with dimensions [d_1, ..., d_n, in_1, ..., in_x], the output dimensions will be [d_1, ..., d_n, out_1, ..., out_y]. The behavior stays consistent with torch.matmul. If out_features=[], output dimensions will be [d_1, ..., d_n] instead of [d_1, ..., d_n, 1].

This is pretty difficult to understand.  This doesn't define whether you are referring to current behavior or new behavior.  Terms are also not defined or are used inconsistently (`in` is used before being defined, `out_features` presumably refers to the output analog of `in` but isn't called `out` (to disambiguate it from an `out=` parameter?).  How about an example?