@gchanan 2019-05-08
> I would vote for this one as benchmarks show no regression.

You haven't said what the alternative is yet.